refactor(mysql): centralize XML sql_mode probe
