+++ src/infra/adapters/mysql/cli/process/adhoc.rs
@@ -14,7 +14,6 @@ use crate::domain::{
     },
 };
 
-use super::super::error::validate_mode_probe;
 use super::super::policy::{
     MySqlExecutionResult, mysql_command_tag, mysql_metadata_fallback_has_unsupported_session_state,
     mysql_possible_refresh_scope, mysql_refresh_scope, mysql_row_count_marker,
@@ -24,8 +23,8 @@ use super::super::xml::MySqlResultSet;
 use super::metadata::mysql_metadata_columns_with_diagnostics;
 use super::{
     MYSQL_QUERY_TIMEOUT, MySqlProcess, configure_mysql_session, finish_mysql_session,
-    read_one_mysql_resultset, read_one_mysql_resultset_with_diagnostics,
-    run_mysql_process_with_timeout, validate_mysql_session_exit, write_mysql_statement,
+    read_one_mysql_resultset_with_diagnostics, run_mysql_process_with_timeout,
+    validate_mysql_session_exit, write_mysql_statement,
 };
 
 pub(in crate::adapters::mysql) async fn run_mysql_adhoc(
@@ -237,15 +236,7 @@ async fn run_mysql_adhoc_process(
     expected_columns: Option<&[&str]>,
 ) -> Result<MySqlExecutionResult, DbOperationError> {
     configure_mysql_session(process, access_mode).await?;
-
-    let probe_marker = Uuid::new_v4().simple().to_string();
-    let probe_query = format!(
-        "SELECT '{probe_marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode"
-    );
-    write_mysql_statement(process, &probe_query).await?;
-    let probe_xml = read_one_mysql_resultset(process).await?;
-    let probe = super::parse_mysql_xml(&probe_xml)?;
-    validate_mode_probe(&probe, &probe_marker)?;
+    process.probe_sql_mode().await?;
 
     let mut last_result_set = None;
     let mut last_result_statement = None;
