+++ src/infra/adapters/mysql/cli/process/single.rs
@@ -4,15 +4,14 @@ use uuid::Uuid;
 use crate::app::ports::outbound::{AccessMode, DbOperationError};
 use crate::domain::RefreshScope;
 
-use super::super::error::validate_mode_probe;
 use super::super::policy::{
     MYSQL_SESSION_MARKER_COLUMN, MySqlExecutionResult, validate_mysql_session_marker,
 };
 use super::super::xml::parse_mysql_xml;
 use super::{
     MYSQL_QUERY_TIMEOUT, MySqlProcess, configure_mysql_session, finish_mysql_session,
-    read_one_mysql_resultset, read_one_mysql_resultset_with_diagnostics,
-    run_mysql_process_with_timeout, validate_mysql_session_exit, write_mysql_statement,
+    read_one_mysql_resultset_with_diagnostics, run_mysql_process_with_timeout,
+    validate_mysql_session_exit, write_mysql_statement,
 };
 
 pub(in crate::adapters::mysql) async fn run_mysql_single_statement(
@@ -38,15 +37,7 @@ pub(super) async fn run_mysql_single_statement_process_with_diagnostics(
     access_mode: AccessMode,
 ) -> Result<MySqlExecutionResult, DbOperationError> {
     configure_mysql_session(process, access_mode).await?;
-
-    let probe_marker = Uuid::new_v4().simple().to_string();
-    let probe_query = format!(
-        "SELECT '{probe_marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode"
-    );
-    write_mysql_statement(process, &probe_query).await?;
-    let probe_xml = read_one_mysql_resultset(process).await?;
-    let probe = parse_mysql_xml(&probe_xml)?;
-    validate_mode_probe(&probe, &probe_marker)?;
+    process.probe_sql_mode().await?;
 
     write_mysql_statement(process, query).await?;
     let (stdout, mut diagnostics) = read_one_mysql_resultset_with_diagnostics(process).await?;
@@ -78,6 +69,7 @@ pub(super) async fn run_mysql_single_statement_process_with_diagnostics(
 #[cfg(test)]
 pub(super) mod test_support {
     use super::super::super::xml::MySqlResultSet;
+    use super::super::read_one_mysql_resultset;
     use super::*;
 
     pub async fn run_mysql_single_statement_process(
@@ -86,14 +78,7 @@ pub(super) mod test_support {
         access_mode: AccessMode,
     ) -> Result<MySqlResultSet, DbOperationError> {
         configure_mysql_session(process, access_mode).await?;
-
-        let marker = Uuid::new_v4().simple().to_string();
-        let probe_query =
-            format!("SELECT '{marker}' AS __sabiql_probe, @@SESSION.sql_mode AS __sabiql_sql_mode");
-        write_mysql_statement(process, &probe_query).await?;
-        let probe_xml = read_one_mysql_resultset(process).await?;
-        let probe = parse_mysql_xml(&probe_xml)?;
-        validate_mode_probe(&probe, &marker)?;
+        process.probe_sql_mode().await?;
 
         write_mysql_statement(process, query).await?;
